Refactoring continues (Completed for real this time?)
Fixed a few small typos in accessing JSON files

Need to build the lower portions of the UI to really make much more progress in manual testing. Need to get either docker or a private file and parser set up for automated testing. Want to do one of the two today